Don't flash alert icons for packed combinators
When a filter combinator is in a packed compakt circuit and the circuit
is now powered, the combinator will flash an alert icon which overlaps
with the "main" alert icon of the cc. There is no need to do so, so this
change turns off the icons for the packed filter combinator.